Bug #43526: only remove VBS files otherwise the icon for teacher's UMC link is always removed
git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-4.1/ucs-school-4.1r2@77359 1b283449-9f2f-0410-b571-8eb228e1a901

+++ ucs-school-netlogon-user-logonscripts/ucs-school-user-logonscript.py
@@ -673,7 +673,8 @@ def clean():
 		for path in UserLogonScriptListener.get_script_path():
 			if os.path.exists(path):
 				for f in os.listdir(path):
-					os.unlink(os.path.join(path, f))
+					if f.lower.endswith('.vbs'):
+						os.unlink(os.path.join(path, f))
 	finally:
 		listener.unsetuid()
